requirements.txt
pandas
fastapi
uvicorn
pydantic
PyYAML
